requirements.txt
google-generativeai==0.3.1
pypdf
pymupdf
python-dotenv